requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
sentence-transformers==3.2.1
chromadb==0.5.23
PyPDF2==3.0.1
requests==2.32.3
numpy>=1.21.0
torch>=2.0.0
python-dotenv>=1.0.0